+++ assets/js/app.js
@@ -60,6 +60,19 @@ Hooks.AssocListChange = {
   },
 };
 
+Hooks.Copy = {
+  mounted() {
+    let { to } = this.el.dataset;
+    this.el.addEventListener('click', ev => {
+      ev.preventDefault();
+      let text = document.querySelector(to).value;
+      navigator.clipboard.writeText(text).then(() => {
+        console.log('Copied!');
+      });
+    });
+  },
+};
+
 // @ts-ignore
 let csrfToken = document
   .querySelector("meta[name='csrf-token']")