+++ lib/lightning/accounts.ex
@@ -142,6 +142,22 @@ defmodule Lightning.Accounts do
   def get_user!(id), do: Repo.get!(User, id)
 
   @doc """
+  Gets a single token.
+
+  Raises `Ecto.NoResultsError` if the UserToken does not exist.
+
+  ## Examples
+
+      iex> get_token!(123)
+      %UserToken{}
+
+      iex> get_token!(456)
+      ** (Ecto.NoResultsError)
+
+  """
+  def get_token!(id), do: Repo.get!(UserToken, id)
+
+  @doc """
   Registers a superuser.
 
   ## Examples
@@ -543,6 +559,30 @@ defmodule Lightning.Accounts do
   def get_user_by_api_token(token) do
     {:ok, query} = UserToken.verify_token_query(token, "api")
     Repo.one(query)
+  end
+
+  @doc """
+  Deletes a token.
+
+  ## Examples
+
+      iex> delete_token(token)
+      {:ok, %UserToken{}}
+
+      iex> delete_token(token)
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_token(%UserToken{} = token) do
+    Repo.delete(token)
+  end
+
+  @doc """
+  Lists all user tokens
+  """
+  def list_api_tokens(user) do
+    UserToken.user_and_contexts_query(user, :api)
+    |> Repo.all()
   end
 
   ## Confirmation