+++ lib/lightning_web/controllers/user_auth.ex
@@ -119,14 +119,21 @@ defmodule LightningWeb.UserAuth do
     end
   end
 
+  defp update_last_used(token) do
+    Lightning.Accounts.UserToken.token_and_context_query(token, "api")
+    |> Lightning.Repo.one()
+    |> Lightning.Accounts.UserToken.last_used_changeset()
+    |> Lightning.Repo.update!()
+  end
+
   def authenticate_bearer(conn, _opts) do
     with {:ok, bearer_token} <- get_bearer(conn),
          user when not is_nil(user) <-
            Accounts.get_user_by_api_token(bearer_token) do
+      update_last_used(bearer_token)
       assign(conn, :current_user, user)
     else
-      {:error, _error} ->
-        conn
+      _ -> conn
     end
   end