+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to
 
 ### Added
 
+- Added ability to create and revoke personal API tokens
+  [#147](https://github.com/OpenFn/Lightning/issues/147)
+- Add `last-used at` to API tokens
+  [#722](https://github.com/OpenFn/Lightning/issues/722)
+
 ### Changed
 
 ### Fixed